feat(sql): Implement ALTER TABLE functionality for column operations
- Add comprehensive ALTER TABLE support for ADD COLUMN and DROP COLUMN operations
- Parser support for ALTER TABLE syntax with optional COLUMN keyword
- AST nodes for AlterTableStmt with AlterTableAction enum types
- Planner integration with LogicalAlterTableAddColumn and LogicalAlterTableDropColumn plans
- Executor operators for both ADD and DROP COLUMN operations
- Extend catalog interface with AddColumn and DropColumn methods
- MemoryCatalog implementation with proper column management
- Support for NOT NULL constraints and other column constraints
- Proper error handling for edge cases (duplicate columns, non-existent tables, last column)
- Comprehensive test coverage for parser, planner, executor, and integration
- Full pipeline testing from SQL parsing to catalog updates

+++ internal/network/extended_query_test.go
@@ -179,7 +179,7 @@ func TestExtendedQueryProtocol(t *testing.T) {
 		if len(retrieved.ParamValues) != 1 {
 			t.Errorf("Expected 1 parameter value, got %d", len(retrieved.ParamValues))
 		}
-		if retrieved.ParamValues[0].Data != int64(123) {
+		if retrieved.ParamValues[0].Data != int32(123) {
 			t.Errorf("Expected parameter value 123, got %v", retrieved.ParamValues[0].Data)
 		}
 	})
@@ -346,7 +346,7 @@ func TestParameterParsing(t *testing.T) {
 			data:     []byte("42"),
 			dataType: types.Integer,
 			format:   0,
-			expected: int64(42),
+			expected: int32(42),
 		},
 		{
 			name:     "String text format",
@@ -388,7 +388,7 @@ func TestParameterParsing(t *testing.T) {
 			data:     []byte("123"),
 			dataType: types.Unknown,
 			format:   0,
-			expected: int64(123),
+			expected: int32(123),
 		},
 		{
 			name:     "Unknown type infers string",
